examples/requirements.txt
langchain-mcp-adapters
langgraph
langchain[openai]
python-dotenv
